+++ com.unity.render-pipelines.high-definition/Runtime/Lighting/LightLoop/LightLoop.cs
@@ -801,13 +801,14 @@ namespace UnityEngine.Experimental.Rendering.HDPipeline
                 return false;
 
             // Ratio of the size of the light on screen and its intensity, gives a value used to compare light importance
-            float lightDominanceValue = light.screenRect.size.magnitude * lightComponent.intensity;
+            float lightDominanceValue = light.screenRect.size.magnitude * lightComponent.intensity;;
 
-            if (additionalShadowData == null || !additionalShadowData.contactShadows || lightComponent.shadows == LightShadows.None)
+            if (additionalShadowData == null || !additionalShadowData.contactShadows)
                 return false;
             if (lightDominanceValue <= m_DominantLightValue || m_DominantLightValue == Single.PositiveInfinity)
                 return false;
 
+            // Directional lights are always considered first.
             if (light.lightType == LightType.Directional)
                 m_DominantLightValue = Single.PositiveInfinity;
             else
@@ -918,7 +919,7 @@ namespace UnityEngine.Experimental.Rendering.HDPipeline
 
             lightData.contactShadowIndex = -1;
 
-            // The first shadow casting directional light with contact shadow enabled is always taken as dominant light
+            // The first directional light with contact shadow enabled is always taken as dominant light
             if (GetDominantLightWithShadows(additionalShadowData, light, lightComponent))
                 lightData.contactShadowIndex = 0;